Fix tooltip displaying entities like &, <, >
The fix is complicated by the fact that activity names which will link
to the activity and labels are treated differently. Using just s() will
fix activities but not labels.
To test you need to have both an activity and a label with all of the
characters you wish to test such as: A & B > C < "D".